fix(m66-001): integration test failures from §1+§2+§3 batch
Three failures surfaced under `make test-integration` after reseed:

1. signup_bootstrap_test: balance assertion still pinned the old 500-cent
   value. Switched to `tenant_billing.STARTER_CREDIT_NANOS` constant per
   Captain's "no hardcoded numerics" rule.

2. tenant_billing_integration_test: balanceCoversEstimate test drained
   only 500 nanos from a 5e9-nanos balance — far above
   STAGE_SELF_MANAGED_NANOS, so the gate didn't block as expected. Drains
   the full STARTER_CREDIT_NANOS now.

3. model_caps handler + integration test: ModelCap struct still typed the
   per-token rate fields as i32 and the SQL row.get() reads as i32, but
   schema/019 widened the columns to BIGINT (nanos values exceed
   INT32_MAX for $30/M tokens). Field types and reads now i64. Test
   assertion updated to expect the nanos-scaled values
   ("input_nanos_per_mtok":3000000000 for sonnet, was 300 in cents).

The fourth failure (svix_signature_test boundary case) is unrelated to
this batch — pre-existing harness flake; tracking separately.

Verified: `make test` (29 unit) and `make test-integration` (1508 tests
incl. real-DB suite) both green.

+++ src/http/handlers/model_caps.zig
@@ -43,8 +43,8 @@ pub const MODEL_CAPS_PATH = "/_um/" ++ MODEL_CAPS_PATH_KEY ++ "/model-caps.json"
 const ModelCap = struct {
     id: []const u8,
     context_cap_tokens: i32,
-    input_nanos_per_mtok: i32,
-    output_nanos_per_mtok: i32,
+    input_nanos_per_mtok: i64,
+    output_nanos_per_mtok: i64,
 };
 
 const ResponseBody = struct {
@@ -138,8 +138,8 @@ fn appendRow(
 ) !void {
     const id = try alloc.dupe(u8, try row.get([]const u8, 0));
     const cap = try row.get(i32, 1);
-    const in_rate = try row.get(i32, 2);
-    const out_rate = try row.get(i32, 3);
+    const in_rate = try row.get(i64, 2);
+    const out_rate = try row.get(i64, 3);
     const updated = try row.get(i64, 4);
     try models.append(alloc, .{
         .id = id,
